backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1

# Async HTTP Client
aiohttp==3.11.11
httpx==0.28.1

# Google AI / Gemini
google-generativeai==0.8.3

# Database (Phase 2)
# sqlalchemy[asyncio]==2.0.36
# asyncpg==0.30.0
# alembic==1.14.0

# Caching (Phase 2)
# redis==5.2.1

# CORS and Security
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1

